requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
numpy>=1.26,<2.1
librosa>=0.10,<0.11
torch>=2.2,<3
python-dotenv>=1.0,<2
pymysql>=1.1,<2
google-cloud-storage>=2.16,<3
fasttext-wheel>=0.9.2,<1
scikit-learn>=1.4,<2
soundfile>=0.12,<1
